Refactor NoteApp and add NoteModel class
